board: phytec: mx7d_phyboard_zeta: pinmux UART based upon CONFIG_CONS_INDEX
This patch allows for pinmuxing of the UART interface to be configured
by CONFIG_CONS_INDEX found in mx7d_phyboard_zeta.h so that U-Boot
displays output to the desired port. If CONFIG_CONS_INDEX is undefined
or points to a UART that has no defined pinmux, we default to UART5.

Signed-off-by: Matt McKee <mmckee@phytec.com>