Do not execute lcdif_power_down() body; data abort issue with M4 running?!
As for now, weird behaviour where this function will cause a data abort,
but only if the M4 is running pingpong_rpmsg_freertos or similar code.
(Root cause to be investigated.)

The data abort handler will reset the CPU, however, this causes a loop,
because the lcdif_power_down() is called inside the CPU reset function.

Unsure what this will solve.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>