Keep i.MX watchdog power down timer running until watchdog starts.
Upon reset the i.MX hardware watchdog always starts its power down
timer, which ensures the i.MX is reset if it gets stuck during early
boot.

However the power down timer got disabled early during start-up,
even before the (normal) i.MX watchdog is enabled. This provides an
unprotected gap where any system glitch can leave the system hanging.

If the (regular) i.MX watchdog is enabled, we want the power down timer
to be disabled only after the (regular) i.MX watchdog has started, so
that the complete U-Boot boot sequence is covered by a hardware watchdog.

The assumption is that the (normal) i.MX watchdog is enabled within 16
seconds after boot; this is a safe assumption as the watchdog is
typically enabled in the board_late() function, which we think is not
*that* late :)

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>